Changed default tokenizer behavior!
Now upon seeing a reader macro, the tokenizer will not push/pop the
indented character stream itself, but will rather delegate the reading
completely to the macro's tokenizer, and let *it* do the push/popping.

This is more general behavior, and allowed for simple changes to the
string tokenizer to result in a block string tokenizer.